Updated User model attribute, unit tests

+++ src/main/java/com/mvidyn/std/phub/ui/repository/OnlineCbftRepository.java
@@ -3,5 +3,5 @@ package com.mvidyn.std.phub.ui.repository;
 import com.mvidyn.std.phub.ui.model.form.online.data.OnlineCbftData;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OnlineCbftRepository extends JpaRepository<OnlineCbftData, Integer> {
+public interface OnlineCbftRepository extends JpaRepository<OnlineCbftData, Long> {
 }
